Refactor sidebar navigation and remove unused pages
- Updated sidebar navigation to reflect new routes and removed the "Upgrade to PRO" link.
- Changed the dashboard route to point to the customer dashboard.
- Removed the "User Management" link and added "My Policies" link in the sidebar.
- Deleted the upgrade page view and its associated route.
- Cleaned up the user profile edit view by removing unnecessary elements.
- Deleted the user management index view.
- Created new views for customer dashboard and my policies.
- Added a new model and migration for approved policies.
