+++ SafeNest/app/Models/ApprovedPolicy.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class ApprovedPolicy extends Model
+{
+    use HasFactory;
+}
