+++ SafeNest/resources/views/customer/dashboard.blade.php
@@ -1,0 +1,13 @@
+@extends('layouts.customerapp', ['activePage' => 'Dashboard', 'title' => 'Sample', 'navName' => 'Customer Dashboard', 'activeButton' => 'laravel'])
+
+@section('content')
+    <div class="content">
+        <div class="container-fluid">
+            
+        </div>
+    </div>
+@endsection
+
+@push('js')
+    
+@endpush
