+++ SafeNest/resources/views/customer/mypolicy.blade.php
@@ -1,0 +1,13 @@
+@extends('layouts.customerapp', ['activePage' => 'My policies', 'title' => 'My policies', 'navName' => 'My Policies', 'activeButton' => 'laravel'])
+
+@section('content')
+    <div class="content">
+        <div class="container-fluid">
+            
+        </div>
+    </div>
+@endsection
+
+@push('js')
+    
+@endpush
