+++ SafeNest/resources/views/layouts/footer/nav.blade.php
@@ -3,16 +3,13 @@
         <nav>
             <ul class="footer-menu">
                 <li>
-                    <a href="https://www.creative-tim.com" class="nav-link" target="_blank">{{ __('Creative Tim') }}</a>
+                    <a href="https://www.creative-tim.com" class="nav-link" target="_blank">{{ __('Fantastic Five') }}</a>
                 </li>
                 <li>
-                    <a href="https://www.updivision.com" class="nav-link" target="_blank">{{ __('Updivision') }}</a>
+                    <a href="https://www.updivision.com" class="nav-link" target="_blank">{{ __('Github') }}</a>
                 </li>
                 <li>
                     <a href="https://www.creative-tim.com/presentation" class="nav-link" target="_blank">{{ __('About Us') }}</a>
-                </li>
-                <li>
-                    <a href="http://blog.creative-tim.com" class="nav-link" target="_blank">{{ __('Blog') }}</a>
                 </li>
             </ul>
             <p class="copyright text-center">
@@ -20,7 +17,7 @@
                 <script>
                     document.write(new Date().getFullYear())
                 </script>
-                <a href="http://www.creative-tim.com">{{ __('Creative Tim') }}</a> &amp; <a href="https://www.updivision.com">{{ __('Updivision') }}</a> {{ __(', made with love for a better web') }}
+                <a href="http://www.creative-tim.com">{{ __('GROUP J') }}</a> &amp; <a href="https://www.updivision.com">{{ __('SONS') }}</a> {{ __(', made due to pressure of deadlines') }}
             </p>
         </nav>
     </div>
