+++ SafeNest/routes/web.php
@@ -17,6 +17,12 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', function () {
     return view('welcome');
 });
+Route::get('/customerdashboard', function () {
+    return view('customer.dashboard');
+})->name('customerdashboard');
+Route::get('/mypolicies', function () {
+    return view('customer.mypolicy');
+})->name('mypolicies');
 
 Route::get('/dashboard', function () {
     return view('dashboard');
